refactor(broker): adjust paths of query api tests
